% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clearwater_lake_density}
\alias{clearwater_lake_density}
\title{Example species density}
\format{
An \code{sf} object with 272 features and 1 variable "density".
}
\usage{
clearwater_lake_density
}
\description{
A dataset containing simulated species density near Clearwater Lake, Manitoba.
}
\details{
Simulated species density with three levels "High", "Medium", Low" with
probabilities 0.1, 0.3, 0.6. Grid size is 2 km x 2 km.
}
\examples{
library(sf)
data(clearwater_lake_density)
plot(clearwater_lake_density)
}
\keyword{datasets}
