% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{clearwater_lake_elevation}
\alias{clearwater_lake_elevation}
\title{Example elevation}
\format{
An external tif file to be read in with {terra} as a \code{SpatRaster} object.
}
\source{
Hollister, J.W. (2022). elevatr: Access Elevation Data from Various APIs. R package version 0.4.2. \url{https://CRAN.R-project.org/package=elevatr/}
}
\description{
A dataset containing elevation near Clearwater Lake, Manitoba.
}
\details{
Elevation data are from the AWS using the {elevatr} package.

The units are meters.
}
\examples{
library(terra)
clearwater_elev_path <- system.file("extdata", "clearwater_lake_elevation.tif", package = "camtrapmonitoring")
clearwater_lake_elevation <- rast(clearwater_elev_path)
plot(clearwater_lake_elevation)
}
