% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clearwater_lake_extent}
\alias{clearwater_lake_extent}
\title{Example extent}
\format{
An \code{sf} object with 2 points.
}
\usage{
clearwater_lake_extent
}
\description{
A dataset containing the extent near Clearwater Lake, Manitoba.
}
\examples{
data(clearwater_lake_extent)
plot(clearwater_lake_extent)
}
\keyword{datasets}
