% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clearwater_lake_roads}
\alias{clearwater_lake_roads}
\title{Example roads}
\format{
An \code{sf} object with 172 roads and 48 variables. See the source below
for details.
}
\source{
\url{https://open.canada.ca/data/en/dataset/3d282116-e556-400c-9306-ca1a3cada77f}
}
\usage{
clearwater_lake_roads
}
\description{
A dataset containing roads near Clearwater Lake, Manitoba.
}
\details{
Roads are from the Canadian National Road Network (NRN) with all
corresponding fields.
}
\examples{
library(sf)
data(clearwater_lake_roads)
plot(clearwater_lake_roads)
}
\keyword{datasets}
