% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clearwater_lake_wetlands}
\alias{clearwater_lake_wetlands}
\title{Example wetland features}
\format{
An \code{sf} object with 5 features and 8 fields. See the source
below for details.
}
\source{
Mark Padgham, Bob Rudis, Robin Lovelace, Maëlle Salmon (2017).
“osmdata.” \emph{Journal of Open Source Software}, \emph{2}(14), 305.
doi:10.21105/joss.00305 \url{https://doi.org/10.21105/joss.00305},
\url{https://joss.theoj.org/papers/10.21105/joss.00305}.
}
\usage{
clearwater_lake_wetlands
}
\description{
A dataset containing wetland features near Clearwater Lake, Manitoba.
}
\details{
Hydrology features are from Open Street Map, downloaded using the {osmdata}
package.
}
\examples{
data(clearwater_lake_wetlands)
plot(clearwater_lake_wetlands)
}
\keyword{datasets}
