% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-to.R
\name{dist_to}
\alias{dist_to}
\title{Distance-to}
\usage{
dist_to(x, y)
}
\arguments{
\item{x}{object of class sfg, sfc or sf}

\item{y}{object of class sfg, sfc or sf}
}
\value{
Vector of distances between x and the nearest feature in y.
}
\description{
Calculate the distance from features in x to the nearest feature in y.
}
\details{
To avoid the large overhead of creating distance to rasters for small/medium number of sample points, this vector-based distance to determines the nearest feature (y) to each x then calculates the distance between each pair.
}
\examples{
data(water)
data(points)

dist_to(points, water)
}
