% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_buffer}
\alias{eval_buffer}
\alias{eval_buffer.data.table}
\alias{eval_buffer,}
\alias{eval_buffer-data.table-method}
\alias{eval_buffer.sf}
\alias{eval_buffer-sf-method}
\title{Eval layers by buffered}
\usage{
eval_buffer(x, layer, buffersize, type, direction, ...)

\method{eval_buffer}{data.table}(x, layer, buffersize, type, direction,
  coords)

\method{eval_buffer}{sf}(x, layer, buffersize, type, direction)
}
\arguments{
\item{x}{data.table or sf points.}

\item{layer}{raster layer.}

\item{buffersize}{radius}

\item{type}{one of 'categorical', 'binary', 'ordinal', or 'real'. See Details.}

\item{direction}{one of 'positive', 'neutral', 'negative'. See Details.}

\item{...}{}

\item{coords}{names of coordinate columns.}
}
\description{
Eval layers by buffered
}
\seealso{
eval_buffer
}
