% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_buffer}
\alias{eval_buffer}
\title{Evaluate camera trap locations by buffered sampling of layers}
\usage{
eval_buffer(
  x,
  y,
  buffer_size,
  buffer_fun = mean,
  layer = 1,
  type,
  direction,
  coords = NULL
)
}
\arguments{
\item{x}{SpatRaster layer (terra package)}

\item{y}{spatial feature object}

\item{layer}{default 1, see terra::extract}

\item{type}{one of 'categorical', 'binary', 'ordinal', or 'real'. See Details.}

\item{direction}{one of 'positive', 'neutral', 'negative'. See Details.}

\item{buffersize}{radius of buffer around each point.}
}
\description{
Using the buffered points locations generated manually or with \code{camtrapmonitoring} functions \code{\link[=sample_ct]{sample_ct()}} and \code{\link[=grid_ct]{grid_ct()}}, sample raster layers to characterize and select camera trap locations, and quantify potential sampling bias.
}
\details{
Type is used to define the data type of the sampled raster layer:
\itemize{
\item 'categorical' - one of a limited and usually fixed number of possible values (e.g.: landcover. 1, 2, 3 indicating 'wetland', 'forest', 'agriculture').
\item 'binary' - also boolean. TRUE/FALSE, presence/absence. (e.g.: raster of wetlands).
\item 'ordinal' - ranked or ordered data. (e.g.: animal density high/medium/low).
\item 'real' - real number variable either interval or ratio scale. (e.g.: digital elevation model).
}

The function used to summarize the buffer regions around each point in \code{x} depends on the data \code{type} defined. 'categorical' data should likely be transformed into individual binary layers with \code{\link[=binary_layer]{binary_layer()}} so \code{mean} can be used to determine proportions in a buffer region.

Direction is used to define the user's priority or preference for this layer. For example, 'positive' direction when evaluating an elevation layer would indicate that the sampled camera trap locations are preferably in areas of high elevation. This attribute is later used by \code{\link[=select_ct]{select_ct()}} for ranking and selecting camera trap locations.
}
\examples{
library(terra)
data("clearwater_lake_density")
clearwater_lake_elevation <- rast(system.file('extdata', 'clearwater_lake_elevation.tif', package = 'camtrapmonitoring'))

# Sample points
pts <- sample_ct(clearwater_lake_density, 1, type = 'random')

# Make grid with queen's case
queen <- grid_ct(pts, case = 'queen', distance = 100)

# Evaluate each point with the land cover layer
queen$elev <- eval_buffer(x = clearwater_lake_elevation, y = queen, buffer_size = 150)

plot(queen["elev"])
}
\seealso{
Other eval: 
\code{\link{eval_dist}()},
\code{\link{eval_pt}()}
}
\concept{eval}
