% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_dist}
\alias{eval_dist}
\title{Evaluate distance-to}
\usage{
eval_dist(x, y)
}
\arguments{
\item{x}{data.table or sf points.}

\item{y}{object of class sfg, sfc or sf.}
}
\value{
Vector of distances between x and the nearest feature in y.
}
\description{
Evaluates locations in x by measuring the distance to the nearest feature in y.
}
\details{
To avoid the large overhead of creating distance to rasters for small/medium number of sample points, this vector-based distance to determines the nearest feature (y) to each x then calculates the distance between each pair.
}
\examples{
data(water)
data(points)

points$distWater <- eval_dist(points, water)
}
\seealso{
Other eval: \code{\link{eval_buffer}},
  \code{\link{eval_pt}}
}
\concept{eval}
