% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_pt}
\alias{eval_pt}
\alias{eval_pt.data.table}
\alias{eval_pt,}
\alias{eval_pt-data.table-method}
\alias{eval_pt.sf}
\alias{eval_pt-sf-method}
\title{Evaluate camera trap locations by point sampling layers}
\usage{
eval_pt(x, layer, type = NULL, direction = NULL, ...)

\method{eval_pt}{data.table}(x, layer, type, direction, coords)

\method{eval_pt}{sf}(x, layer, type, direction)
}
\arguments{
\item{x}{data.table or sf points.}

\item{layer}{raster layer.}

\item{type}{one of 'categorical', 'binary', 'ordinal', or 'real'. See Details.}

\item{direction}{one of 'positive', 'neutral', 'negative'. See Details.}

\item{...}{}

\item{coords}{names of coordinate columns.}
}
\description{
Using the point locations generated manually or with \code{wildcam} functions \code{\link[=strat_sample]{strat_sample()}} and \code{\link[=make_grid]{make_grid()}}, sample raster layers to characterize and select camera trap locations and understand sampling bias
}
\examples{
# Load data
data(points)
data(lc)

# Evaluate each point with the land cover layer
#   type is categorical, and the direction is neutral
points$lc <- eval_pt(x = points, layer = lc, type = 'categorical', direction = 'neutral')

plot(points)
}
\seealso{
eval_buffer
}
