% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_pt}
\alias{eval_pt}
\title{Evaluate camera trap locations by point sampling layers}
\usage{
eval_pt(x, y, layer = 1, type = NULL, direction = NULL, coords = NULL)
}
\arguments{
\item{x}{SpatRaster layer (terra package)}

\item{y}{spatial feature object}

\item{layer}{default 1, see terra::extract}

\item{type}{one of 'categorical', 'binary', 'ordinal', or 'real'. See Details.}

\item{direction}{one of 'positive', 'neutral', 'negative'. See Details.}
}
\description{
Using the point locations generated manually or with \code{camtrapmonitoring} functions \code{\link[=sample_ct]{sample_ct()}} and \code{\link[=grid_ct]{grid_ct()}}, sample raster layers to characterize and select camera trap locations, and quantify potential sampling bias.
}
\details{
Type is used to define the data type of the sampled raster layer:
\itemize{
\item 'categorical' - one of a limited and usually fixed number of possible values (e.g.: landcover. 1, 2, 3 indicating 'wetland', 'forest', 'agriculture').
\item 'binary' - also boolean. TRUE/FALSE, presence/absence. (e.g.: raster of wetlands).
\item 'ordinal' - ranked or ordered data. (e.g.: animal density high/medium/low).
\item 'real' - real number variable either interval or ratio scale. (e.g.: digital elevation model).
}

Direction is used to define the user's priority or preference for this layer. For example, 'positive' direction when evaluating an elevation layer would indicate that the sampled camera trap locations are preferably in areas of high elevation. This attribute is later used by \code{\link[=select_ct]{select_ct()}} for ranking and selecting camera trap locations.
}
\examples{
data("clearwater_lake_density")
clearwater_lake_land_cover <- rast(system.file('extdata', 'clearwater_lake_land_cover.tif', package = 'camtrapmonitoring'))

# Sample points
pts <- sample_ct(clearwater_lake_density, 1, type = 'random')

# Make grid with queen's case
queen <- grid_ct(pts, case = 'queen', distance = 100)

# Evaluate each point with the land cover layer
queen$lc <- eval_pt(x = points, layer = clearwater_lake_land_cover, type = 'categorical', direction = 'neutral')

plot(queen["lc"])
}
\seealso{
Other eval: 
\code{\link{eval_buffer}()},
\code{\link{eval_dist}()}
}
\concept{eval}
