% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist-to.R
\name{feat_dist_to}
\alias{feat_dist_to}
\title{Feature-based distance to}
\usage{
feat_dist_to(x, y)
}
\arguments{
\item{x}{object of class sfg, sfc or sf}

\item{y}{object of class sfg, sfc or sf}
}
\value{
Vector of distances between x and the nearest feature in y.
}
\description{
To avoid the large overhead of creating distance to rasters for small/medium number of sample points, this "vector-based distance to" determines the nearest feature (y) to each x before calculating the distance.
}
