% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-ct.R
\name{grid_design}
\alias{grid_design}
\title{Grid design}
\usage{
grid_design(distance, case, n)
}
\arguments{
\item{distance}{distance between adjacent camera traps. Don't worry about
the hypotenuse.}

\item{case}{"queen", "rook" or "bishop". Ignored if \code{n} is provided.}

\item{n}{number of points around each focal point. \code{n} overrides the \code{case}
argument, do not provide both - see Details.}
}
\description{
Helper function used internally by \code{grid_ct} to establish grids around
focal locations. Provided to the user to explore grid design options before
applying to their data.
}
\examples{
plot(grid_design(distance = 100, case = 'queen'))
plot(grid_design(distance = 100, n = 13))
}
