% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_binary.R
\name{make_binary}
\alias{make_binary}
\title{Binary rasters}
\usage{
make_binary(layer, value)
}
\arguments{
\item{value}{}
}
\value{
A binary raster layer with two values:
}
\description{
Make a binary raster from input 'layer'.
}
