% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\alias{make_grid.data.table}
\alias{make_grid.sf}
\title{Make camera trap grids}
\usage{
make_grid(x, n, case, distance, id = NULL, coords = NULL)

\method{make_grid}{data.table}(x, n, case, distance, id = NULL,
  coords = NULL)

\method{make_grid}{sf}(x, n, case, distance, id = NULL, coords = NULL)
}
\arguments{
\item{x}{data.table or sf points.}

\item{n}{number of points around each focal point. \code{n} overrides the \code{case} argument, do not provide both. See details.}

\item{case}{"queen", "rook" or "bishop". Ignored if \code{n} is provided.}

\item{distance}{distance between adjacent camera traps. Don't worry about the hypotenuse.}

\item{id}{column in \code{x} indicating id of focal point. Only used when x is a \code{data.table}.}

\item{coords}{columns in \code{x} indicating names of coordinate columns of focal point. Only used when x is a \code{data.table}.}
}
\value{
Extended data.table either nine times the length of input DT for 'queen' case or 5 times the length of input DT for 'rook' or 'bishop' case. See examples.

The logical 'focal' column indicates which point is the focal or center camera trap.
}
\description{
Set up grids around focal points. For example, sample points in your study area and use \code{make_grid} to establish a grid of camera traps around each.
}
\examples{
# Point data (sf object)
library(sf)
data(points)
plot(points)

## Make grid with case
queen <- make_grid(points, case = 'queen', distance = 100)
plot(queen)

rook <- make_grid(points, case = 'rook', distance = 100)
plot(rook)

bishop <- make_grid(points, case = 'bishop', distance = 100)
plot(bishop)

## Make grid with n
grid <- make_grid(points, n = 25, distance = 100)
plot(grid)

# data.table input
library(data.table)
DT <- data.table(ID = points$ID, st_coordinates(points))
grid <- make_grid(DT, case = 'queen', distance = 100, id = 'ID', coords = c('X', 'Y'))
}
