% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make camera trap grids}
\usage{
make_grid(x, case, distance, id, coords)
}
\arguments{
\item{case}{"queen", "rook" or "bishop".}

\item{distance}{distance in x and y between cameras. don't worry about the hypotenuse.}

\item{id}{id of focal point.}

\item{coords}{names of coordinate columns.}

\item{DT}{data.table or sf points.}
}
\value{
Extended data.table either nine times the length of input DT for 'queen' case or 5 times the length of input DT for 'rook' or 'bishop' case.

The logical 'focal' column indicates which point is the focal.
}
\description{
Set up grids around focal points. For example, sample points in your study area and establish a grid of camera traps around each.
}
\examples{
grid <- make_grid(DT, id = 'point', case = 'queen', distance = 250, coords = c('X', 'Y'))
# plot 3 types
}
