% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points}
\alias{points}
\title{Example points}
\format{An \code{sf} object with 5 points and two variables:
\itemize{
\item ID: individual identifier, numeric type
\item geometry: \code{sfc} simple feature column
}

The projection for this layer is UTM 14N WGS84. The proj4string for this layer is "+proj=utm +zone=14 +datum=WGS84 +units=m +no_defs", EPSG code 32614.}
\usage{
points
}
\description{
A dataset containing 5 random points nearby three lakes in Manitoba (\code{data(water)}).
}
\examples{
data(points)
plot(points)
}
\keyword{datasets}
