% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-ct.R
\name{sample_ct}
\alias{sample_ct}
\title{Camera trap sampling}
\usage{
sample_ct(x, n, type, col = NULL)
}
\arguments{
\item{x}{spatial feature object}

\item{n}{number of random points, if a column name is provided to \code{col}
n represents the number of random points per strata}

\item{type}{type of sampling, see \link[sf]{st_sample}}

\item{col}{column name in x indicating strata}
}
\value{
\code{sf} object with POINT geometry
}
\description{
Sample potential camera trap locations. For stratified sampling, provide a
suitable column to stratify by. Alternatively, \link[sf]{st_sample} is used
directly to sample points across all features.
}
\examples{
# Example grid with density levels (High, Medium, Low)
data(clearwater_lake_density)

# Stratified random points for each density level
pts_random <- sample_ct(
  x = clearwater_lake_density, n = 20,
  type = 'random', col = 'density')

# Plot density grid and sampled points
plot(clearwater_lake_density, reset = FALSE)
plot(pts_random, add = TRUE, pch = 1)

# Regular sampled points across all features
pts_regular <- sample_ct(
  x = clearwater_lake_density, n = 20, type = 'regular')

# Plot density grid and sampled points
plot(clearwater_lake_density, reset = FALSE)
plot(pts_regular, add = TRUE, pch = 2)
}
