% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_roi}
\alias{scale_roi}
\title{Scale in region of interest}
\usage{
scale_roi(layer, roi)
}
\arguments{
\item{roi}{sf}
}
\description{
Scale a layer in a region of interest to optionally compare locations with [eval_pt()] and [eval_buffer()], and select locations with [select_ct()] based off of relative values instead of absolute values.
}
