% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select-ct.R
\name{select_ct}
\alias{select_ct}
\title{Select camera trap locations}
\usage{
select_ct(x, n, rank, sub, by)
}
\arguments{
\item{sub}{a named list with the form \code{list(colname = value)}. See Details.}

\item{by}{}
}
\description{
...
}
\details{
\code{sub} is a named list used to subset the input \code{x}. It should follow the form \code{list(colname = value)}or \code{list(colname = c(values, values))}.
}
