% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_sample.R
\name{strat_sample}
\alias{strat_sample}
\title{Stratified polygon sampling}
\usage{
strat_sample(x, n, col, returnDT = FALSE)
}
\arguments{
\item{x}{polygon object of class \code{sf}}

\item{n}{number of random points}

\item{col}{column in x indicating strata}

\item{returnDT}{return a \code{data.table} (TRUE) or \code{sf} (FALSE) object}
}
\value{
Either a \code{sf} object or a \code{data.table} with a \code{sfc} (simple feature column).
}
\description{
For each mutually exclusive strata, sample random points.
}
\details{
Polygons cannot be assigned to multiple strata. Strata are defined by values in 'col'. Optionally return a \code{data.table} if 'returnDT' is TRUE or an \code{sf} object if FALSE.
}
\examples{
# Example polygons with density levels 1, 2 and 3.
data(densitygrid)

# Randomly sample 5 points for each set of polygons in each strata.
pts <- strat_sample(x = densitygrid, n = 5, col = 'density', returnDT = FALSE)

plot(densitygrid, reset = FALSE)
plot(pts$geometry, add = TRUE)
}
