% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Water polygons}
\format{An \code{sf} object with 3 polygons and four variables:
\itemize{
\item ID: individual identifier, numeric type
\item PERIMETER: perimeter of the water body
\item AREA: area of the water body
\item geometry: \code{sfc} simple feature column
}}
\source{
\href{http://mli2.gov.mb.ca/roads_hwys/index.html}{Manitoba Land Initiative}
}
\usage{
water
}
\description{
A dataset containing three polygons representing three lakes in Manitoba.
}
\examples{
data(water)
plot(water)
}
\keyword{datasets}
