% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wildcam-pkg.R
\docType{package}
\name{wildcam}
\alias{wildcam}
\alias{_PACKAGE}
\alias{wildcam-package}
\title{wildcam}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Alec L. Robitaille \email{robit.alec@gmail.com} (0000-0002-4706-1762)

}
