% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary-layer.R
\name{binary_layer}
\alias{binary_layer}
\title{Binary rasters}
\usage{
binary_layer(layer, value)
}
\arguments{
\item{layer}{raster layer.}

\item{value}{numeric value in 'layer'. expected length is 1.}
}
\value{
A binary raster layer with two values: \code{TRUE} if pixel matches 'value' provided and \code{FALSE} if pixel does not match 'value' provided.
}
\description{
Make a binary raster from input 'layer'.
}
\details{
Find all pixels matching the 'value' provided and return binary raster.
}
\examples{
data(lc)

bin <- binary_layer(lc, 212)

image(bin)
}
