% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_buffer}
\alias{eval_buffer}
\title{Evaluate camera trap locations by buffered sampling of layers}
\usage{
eval_buffer(
  x,
  y,
  buffer_size,
  buffer_fun = mean,
  layer = 1,
  type,
  direction,
  coords = NULL
)
}
\arguments{
\item{x}{SpatRaster layer (terra package)}

\item{y}{spatial feature object}

\item{layer}{default 1, see terra::extract}

\item{type}{one of 'categorical', 'binary', 'ordinal', or 'real'. See Details.}

\item{direction}{one of 'positive', 'neutral', 'negative'. See Details.}

\item{buffersize}{radius of buffer around each point.}
}
\description{
Using the buffered points locations generated manually or with \code{camtrapmonitoring} functions \code{\link[=sample_ct]{sample_ct()}} and \code{\link[=grid_ct]{grid_ct()}}, sample raster layers to characterize and select camera trap locations, and quantify potential sampling bias.
}
\examples{
library(terra)
data("clearwater_lake_density")
clearwater_lake_elevation <- rast(system.file('extdata', 'clearwater_lake_elevation.tif', package = 'camtrapmonitoring'))

# Sample points
pts <- sample_ct(clearwater_lake_density, 1, type = 'random')

# Make grid with queen's case
queen <- grid_ct(pts, case = 'queen', distance = 100)

# Evaluate each point with the land cover layer
queen$elev <- eval_buffer(x = clearwater_lake_elevation, y = queen, buffer_size = 150)

plot(queen["elev"])
}
\seealso{
Other eval: 
\code{\link{eval_dist}()},
\code{\link{eval_pt}()}
}
\concept{eval}
