% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_dist}
\alias{eval_dist}
\title{Evaluate distance-to}
\usage{
eval_dist(features, target, measure = NULL)
}
\arguments{
\item{features}{sf features}

\item{target}{sf target}

\item{measure}{measure type see geodist::geodist for details}
}
\value{
vector of distances between target and features
}
\description{
Using camera trap locations generated with \code{camtrapmonitoring}
functions \code{\link[=sample_ct]{sample_ct()}} and \code{\link[=grid_ct]{grid_ct()}}, evalaute the distance between
features and camera trap locations to characterize and select locations,
and quantify potential sampling bias.
}
\details{
To avoid the large overhead of creating distance to rasters for small/medium
number of sample points, this vector-based distance to determines the
nearest feature (layer) to each target then calculates the distance between
each pair.
}
\examples{
data("clearwater_lake_density")
data("clearwater_lake_wetlands")

# Sample points
pts <- sample_ct(region = clearwater_lake_density, 1, type = 'random')

# Make grid with queen's case
queen <- grid_ct(features = pts, case = 'queen', distance = 100)

# Evaluate each point with the land cover layer
queen$dist_wetland <- eval_dist(features = queen, target = clearwater_lake_wetlands)

# Plot
plot(queen["dist_wetland"])
}
\seealso{
Other eval: 
\code{\link{eval_buffer}()},
\code{\link{eval_pt}()}
}
\concept{eval}
