% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_dist}
\alias{eval_dist}
\title{Evaluate distance-to}
\usage{
eval_dist(x, y, measure = NULL)
}
\arguments{
\item{x}{SpatRaster layer (terra package)}

\item{y}{spatial feature object}

\item{measure}{measure type see geodist::geodist for details}
}
\value{
vector of distances between x and y

Note attributes are returned like by \code{eval_pt} and \code{eval_buffer}. The
\code{type} attribute for distance to a feature (layer) is "real" and the
\code{direction} is left for the user to provide.
}
\description{
Evaluates locations in x by measuring the distance to the nearest feature
in layer.
}
\details{
To avoid the large overhead of creating distance to rasters for small/medium
number of sample points, this vector-based distance to determines the
nearest feature (layer) to each x then calculates the distance between
each pair.
}
\examples{
data("clearwater_lake_density")
data("clearwater_lake_wetlands")

# Sample points
pts <- sample_ct(clearwater_lake_density, 1, type = 'random')

# Make grid with queen's case
queen <- grid_ct(pts, case = 'queen', distance = 100)

# Evaluate each point with the land cover layer
queen$dist_wetland <- eval_dist(x = clearwater_lake_wetlands, y = queen)

# Plot
plot(queen["dist_wetland"])
}
\seealso{
Other eval: 
\code{\link{eval_buffer}()},
\code{\link{eval_pt}()}
}
\concept{eval}
