% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_pt}
\alias{eval_pt}
\alias{eval_pt.data.table}
\alias{eval_pt,}
\alias{eval_pt-data.table-method}
\alias{eval_pt.sf}
\alias{eval_pt-sf-method}
\title{Evaluate camera trap locations by point sampling layers}
\usage{
eval_pt(x, layer, type = NULL, direction = NULL, ...)

\method{eval_pt}{data.table}(x, layer, buffersize, type, direction, coords)

\method{eval_pt}{sf}(x, layer, buffersize, type, direction)
}
\arguments{
\item{x}{data.table or sf points.}

\item{...}{}

\item{coords}{names of coordinate columns.}
}
\description{
Using the point locations generated manually or with \code{wildcam} functions \code{strat_sample()} and \code{make_grid()}, sample relevant layers to understand sampling bias and assist camera trap location selection.
}
