% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_pt}
\alias{eval_pt}
\title{Evaluate camera trap locations by point sampling layers}
\usage{
eval_pt(features, target, layer = 1)
}
\arguments{
\item{features}{sf features}

\item{target}{SpatRaster target}

\item{layer}{default 1, see terra::extract}
}
\value{
vector of values from target matching locations in features
}
\description{
Using camera trap locations generated with \code{camtrapmonitoring}
functions \code{\link[=sample_ct]{sample_ct()}} and \code{\link[=grid_ct]{grid_ct()}}, sample raster layers to
characterize and select camera trap locations, and quantify potential
sampling bias.
}
\examples{
library(terra)

data("clearwater_lake_density")
clearwater_lake_land_cover <- rast(system.file('extdata',
  'clearwater_lake_land_cover.tif', package = 'camtrapmonitoring'))

# Sample points
pts <- sample_ct(region = clearwater_lake_density, n = 1, type = 'random')

# Make grid with queen's case
queen <- grid_ct(features = pts, case = 'queen', distance = 100)

# Evaluate each point with the land cover layer
queen$lc <- eval_pt(features = queen, target = clearwater_lake_land_cover)

plot(queen["lc"])
}
\seealso{
Other eval: 
\code{\link{eval_buffer}()},
\code{\link{eval_dist}()}
}
\concept{eval}
