% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-ct.R
\name{grid_ct}
\alias{grid_ct}
\title{Make camera trap grids}
\usage{
grid_ct(x, case, distance, id = "id_sample_ct", n)
}
\arguments{
\item{x}{spatial feature object}

\item{case}{"queen", "rook" or "bishop". Ignored if \code{n} is provided.}

\item{distance}{distance between adjacent camera traps. Don't worry about the hypotenuse.}

\item{id}{default: "id_sample_ct" generated automatically from \code{sample_ct}}

\item{n}{number of points around each focal point. \code{n} overrides the \code{case} argument, do not provide both - see Details.}
}
\value{
Extended data.table either nine times the length of input DT for 'queen' case or 5 times the length of input DT for 'rook' or 'bishop' case. See examples.

The logical 'focal' column indicates which point is the focal or center camera trap.
}
\description{
Set up grids around focal points. For example, sample points in your study
area with \code{sample_ct} then use \code{grid_ct} to establish a grid of camera traps
around each.
}
\examples{
data("clearwater_lake_density")
pts <- sample_ct(clearwater_lake_density, 1, type = 'random')

# Make grid with case, eg. 'queen'
queen <- grid_ct(pts, case = 'queen', distance = 100)

# Plot
plot(queen['focal'])

# Make grid with n
n_grid <- grid_ct(pts, n = 25, distance = 100)
plot(n_grid['id_grid_ct'])
}
