% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make camera trap grids}
\usage{
make_grid(x, case, distance, ...)
}
\arguments{
\item{case}{"queen", "rook" or "bishop".}

\item{distance}{distance in x and y between cameras. don't worry about the hypotenuse.}

\item{DT}{data.table or sf points.}

\item{id}{id of focal point.}

\item{coords}{names of coordinate columns.}
}
\value{
Extended data.table either nine times the length of input DT for 'queen' case or 5 times the length of input DT for 'rook' or 'bishop' case.

The logical 'focal' column indicates which point is the focal.
}
\description{
Set up grids around focal points. For example, sample points in your study area and establish a grid of camera traps around each.
}
\examples{
# Point data (sf object)
library(sf)
data(points)
plot(points)

queen <- make_grid(points, case = 'queen', distance = 100)
plot(queen)

rook <- make_grid(points, case = 'rook', distance = 100)
plot(rook)

bishop <- make_grid(points, case = 'bishop', distance = 100)
plot(bishop)

# Or a data.table
library(data.table)
DT <- data.table(ID = points$ID, st_coordinates(points))
make_grid(DT, case = 'queen', distance = 100, id = 'ID', coords = c('X', 'Y'))
}
