% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make camera trap grids}
\usage{
make_grid(x, case, distance, ...)
}
\arguments{
\item{x}{data.table or sf points.}

\item{case}{"queen", "rook" or "bishop".}

\item{distance}{distance between adjacent camera traps. Don't worry about the hypotenuse.}

\item{id}{id of focal point.}

\item{coords}{names of coordinate columns.}
}
\value{
Extended data.table either nine times the length of input DT for 'queen' case or 5 times the length of input DT for 'rook' or 'bishop' case. See examples.

The logical 'focal' column indicates which point is the focal or center camera trap.
}
\description{
Set up grids around focal points. For example, sample points in your study area and use \code{make_grid} to establish a grid of camera traps around each.
}
