% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale-layer.R
\name{scale_layer}
\alias{scale_layer}
\title{Scale in region of interest}
\usage{
scale_layer(x, roi, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{SpatRaster layer (terra package)}

\item{roi}{any object which can be passed to extent including \code{sf}, \code{Spatial}, \code{Raster} objects and 2x2 matrices.}

\item{scale}{see terra::scale}
}
\value{
\code{SpatRaster} layer, cropped to extent of provided 'roi', and scaled.
}
\description{
Scale a layer in a region of interest to optionally compare locations with \code{\link[=eval_pt]{eval_pt()}} and \code{\link[=eval_buffer]{eval_buffer()}}, and select locations based off of relative values instead of absolute values.
}
\examples{
# Load packages
library(terra)

# Load data
clearwater_elev_path <- system.file("extdata", "clearwater_lake_elevation.tif", package = "camtrapmonitoring")
clearwater_lake_elevation <- rast(clearwater_elev_path)

# Region of interest: Clearwater lake area
roi <- ext(clearwater_lake_elevation)

# Scale elevation in extent of density grid
elev_scaled <- scale_layer(clearwater_lake_elevation, roi)

# Plot
plot(elev_scaled)
}
\seealso{
\code{\link[terra:scale]{terra::scale()}}
}
