% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strat_sample.R
\name{strat_sample}
\alias{strat_sample}
\title{Stratified polygon sampling}
\usage{
strat_sample(x, n, col, returnDT = TRUE)
}
\arguments{
\item{x}{polygon object of class sf}

\item{n}{number of random points}

\item{col}{column in x indicating strata}

\item{returnDT}{return a data.table (TRUE) or sf (FALSE) object}
}
\value{
Either a \code{sf} object or a \code{data.table} with a \code{sfc} (simple feature column).
}
\description{
For each unique value in 'col', sample 'n' points and optionally return as a \code{data.table}.
}
