% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water}
\alias{water}
\title{Example water polygons}
\format{An \code{sf} object with 3 polygons and four variables:
\itemize{
\item ID: individual identifier, numeric type
\item PERIMETER: perimeter of the water body
\item AREA: area of the water body
\item geometry: \code{sfc} simple feature column
}

The projection for this layer is UTM 14N WGS84. The proj4string for this layer is "+proj=utm +zone=14 +datum=WGS84 +units=m +no_defs", EPSG code 32614.}
\source{
\href{http://mli2.gov.mb.ca/roads_hwys/index.html}{Manitoba Land Initiative}
}
\usage{
water
}
\description{
A dataset containing three polygons representing three lakes in Manitoba.
}
\examples{
data(water)
plot(water)
}
\keyword{datasets}
